+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.8.6] - UNRELEASED
+## [0.9.0] - 19-06-20
 ### Added
+- Added `jsonKeyEncodingStrategy` to `Request`s, allowing custom encoding strategies for request params.
 - Added `LogEvent.startupInfo`  to keep track of some debugging info while booting.
 
 ### Changed